+++ components/render/OpenGL/scopedbindgl.h
@@ -1,24 +1,29 @@
 #pragma once
+#include <glad/glad.h>
+#include <optional>
 
 namespace Render
 {
     class BindableGL
     {
     public:
-        virtual void bind() = 0;
-        virtual void unbind() = 0;
+        virtual void bind(std::optional<GLenum> binding) = 0;
+        virtual void unbind(std::optional<GLenum> binding) = 0;
     };
 
     class ScopedBindGL
     {
     public:
-        ScopedBindGL(BindableGL& resource) : mResource(resource) { mResource.bind(); }
+        explicit ScopedBindGL(BindableGL& resource, std::optional<GLenum> binding = std::nullopt) : mResource(resource), mBinding(binding)
+        {
+            mResource.bind(mBinding);
+        }
+        explicit ScopedBindGL(BindableGL* resource, std::optional<GLenum> binding = std::nullopt) : ScopedBindGL(*resource, binding) {}
 
-        ScopedBindGL(BindableGL* resource) : ScopedBindGL(*resource) {}
-
-        ~ScopedBindGL() { mResource.unbind(); }
+        ~ScopedBindGL() { mResource.unbind(mBinding); }
 
     private:
         BindableGL& mResource;
+        std::optional<GLenum> mBinding;
     };
 }
